requirements.txt
pandas
numpy
scikit-learn
jupyter
matplotlib
seaborn
fastapi
uvicorn
python-multipart
openpyxl
requests
beautifulsoup4
lxml
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
tiktoken
sentence-transformers
langchain-huggingface
streamlit
